stride-fastapi-azure/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
openai>=1.48.0
python-multipart>=0.0.9
pydantic>=2.8.2
